float issue fix (#692)

+++ uiutils.py
@@ -72,7 +72,7 @@ def getXMLFont(xml, scale = 1):
     if font_italic == '1':
         fnt_flags |= gui.QFont.StyleItalic
 
-    font_size = font_size / float(scale) * 14.
+    font_size = int(font_size / float(scale) * 14.)
     qfnt = gui.QFont(font_name, font_size, fnt_flags)
     qfnt.setPixelSize(font_size)
     return qfnt
